+++ src/mlink/handlers/native.js
@@ -1,23 +1,33 @@
 const mlink = require('../midware');
 const Router = mlink.Router;
 const DeviceManager = require('../managers/device_manager');
-const bundleWrapper = require('../../util/bundle_wrapper');
+const { bundleWrapper, apiWrapper } = require('../../util/wrapper');
 const MemoryFile = require('../../lib/memory_file');
 const debuggerRouter = Router.get('debugger');
-const util = require('../../util');
+const crypto = require('../../util/crypto');
+const path = require('path');
+const LOGLEVEL = {
+  debug: 0,
+  log: 1,
+  info: 2,
+  error: 3,
+  warning: 4,
+  warn: 4
+};
+
 debuggerRouter.registerHandler(function (message) {
   const payload = message.payload;
+  const device = DeviceManager.getDevice(message.channelId);
   if (payload.method === 'WxDebug.initJSRuntime') {
-    const device = DeviceManager.getDevice(message.channelId);
     payload.params.url = new MemoryFile('js-framework.js', payload.params.source).getUrl();
-    if (device.logLevel) {
+    if (device && device.logLevel) {
       payload.params.env.WXEnvironment.logLevel = device.logLevel;
     }
   }
   else if (payload.method === 'WxDebug.callJS' && payload.params.method === 'createInstance') {
     let code = payload.params.args[1];
     if (payload.params.args[2] && (payload.params.args[2]['debuggable'] === 'false' || payload.params.args[2]['debuggable'] === false)) {
-      code = util.obfuscate(code);
+      code = crypto.obfuscate(code);
     }
     debuggerRouter.pushMessageByChannelId('page.debugger', message.channelId, {
       method: 'WxDebug.bundleRendered',
@@ -25,10 +35,37 @@ debuggerRouter.registerHandler(function (message) {
         bundleUrl: payload.params.args[2].bundleUrl
       }
     });
-    payload.params.sourceUrl = new MemoryFile(payload.params.args[2].bundleUrl || (util.md5(code) + '.js'), bundleWrapper(code)).getUrl();
+    payload.params.sourceUrl = new MemoryFile(payload.params.args[2].bundleUrl || (crypto.md5(code) + '.js'), bundleWrapper(code)).getUrl();
+  }
+  else if (payload.method === 'WxDebug.callJS' && payload.params.method === 'createInstanceContext') {
+    if (device.platform === 'iOS') {
+      if (payload.params.args.length < 5) {
+        payload.params.args.splice(1, 0, '');
+      }
+    }
+    const code = payload.params.args[1];
+    const options = payload.params.args[2];
+    const dependenceCode = payload.params.args[4];
+    if (dependenceCode) {
+      payload.params.dependenceUrl = new MemoryFile(`${path.dirname(options.bundleUrl)}/imported_${crypto.md5(dependenceCode)}.js`, apiWrapper(dependenceCode)).getUrl();
+    }
+    else {
+      payload.params.dependenceUrl = '';
+    }
+    if (code) {
+      payload.params.sourceUrl = new MemoryFile(options.bundleUrl || (crypto.md5(code) + '.js'), bundleWrapper(code)).getUrl();
+    }
+    else {
+      payload.params.sourceUrl = '';
+    }
+  }
+  else if (payload.method === 'WxDebug.callJS' && payload.params.method === 'importScript') {
+    const code = payload.params.args[1];
+    const bundleUrl = payload.params.args[2] && payload.params.args[2].bundleUrl;
+    payload.params.sourceUrl = new MemoryFile((bundleUrl || crypto.md5(code) + '.js'), bundleWrapper(code)).getUrl();
   }
   else if (payload.method === 'WxDebug.importScript') {
-    payload.params.sourceUrl = new MemoryFile('imported_' + util.md5(payload.params.source) + '.js', payload.params.source).getUrl();
+    payload.params.sourceUrl = new MemoryFile('imported_' + crypto.md5(payload.params.source) + '.js', payload.params.source).getUrl();
   }
   else if (payload.method === 'syncReturn') {
     message.payload = {
@@ -42,30 +79,45 @@ debuggerRouter.registerHandler(function (message) {
     message.to('page.debugger');
     return;
   }
+  else if (payload.params.method === 'callJS') {
+  }
   else if (payload.method === 'WxDebug.sendSummaryInfo') {
+    message.to('page.debugger');
+    return;
+  }
+  else if (payload.method === 'WxDebug.sendPerformanceData') {
     message.to('page.debugger');
     return;
   }
   message.to('runtime.worker');
 }).at('proxy.native').when('payload.method&&payload.method.split(".")[0]==="WxDebug"');
+
 debuggerRouter.registerHandler(function (message) {
   const payload = message.payload;
+  const device = DeviceManager.getDevice(message.channelId);
   if (payload.method === 'Page.screencastFrame') {
     payload.params.sessionId = 1;
   }
   else if (payload.method === 'Console.messageAdded') {
-    message.payload = {
-      'method': 'Runtime.consoleAPICalled',
-      'params': {
-        'type': payload.params.message.level,
-        'args': payload.params.message.parameters,
-        'executionContextId': 1
+    if (LOGLEVEL[payload.params.message.level] >= LOGLEVEL[device && device.logLevel ? device.logLevel : 'debug']) {
+      message.payload = {
+        'method': 'Runtime.consoleAPICalled',
+        'params': {
+          'type': payload.params.message.level,
+          'args': [{
+            type: 'string',
+            value: payload.params.message.text
+          }] || [],
+          'executionContextId': 1
           // "stackTrace": payload.params.message.stackTrace
-      }
-    };
+        }
+      };
+    }
+    else {
+      message.discard();
+    }
   }
   else if (payload.method === 'DOM.childNodeRemoved') {
-    const device = DeviceManager.getDevice(message.channelId);
     // 此处是为了 扫bundle二维码通知页面关掉bundle二维码界面这个功能
     // 当没有打开JS Debug时 weex加载bundle devtool是不知道的 只能模糊的通过childNodeRemoved判断
     // 这个标记用来限制短时间内发出很多次通知
@@ -90,5 +142,12 @@ debuggerRouter.registerHandler(function (message) {
   else if (payload.result && payload.id === undefined) {
     message.discard();
   }
-  message.to('proxy.inspector');
+  // remove useless but large message
+  if (payload.method && payload.method === 'Page.screencastFrameAck') {
+    message.discard();
+  }
+  else {
+    message.to('proxy.inspector');
+  }
+  // message.to('proxy.inspector');
 }).at('proxy.native').when('!payload.method||payload.method.split(".")[0]!=="WxDebug"');