+++ src/mlink/managers/simulator_manager.js
@@ -1,11 +1,12 @@
-const simrun = require('simrun');
 const config = require('../../lib/config');
 const hook = require('../../util/hook');
+const simrun = require('../../util/simrun');
 exports.connect = function (channelId) {
-  let params = `ws://${config.ip}:${config.port}/debugProxy/native/${channelId}`;
+  let params = '';
   if (config.bundleUrls && config.bundleUrls.length === 1) {
-    params += ',' + config.bundleUrls[0];
+    params += `_wx_tpl=${config.bundleUrls[0]}&`;
   }
+  params += `_wx_devtool=ws://${config.ip}:${config.port}/debugProxy/native/${channelId}`;
   hook.record('/weex_tool.weex_debugger.scenes', { feature: 'simrun', status: 'start' });
-  return simrun.ios('weex-devtool', 'iPhone 6', 'https://registry.npm.taobao.org/weex-devtool-playground/-/weex-devtool-playground-0.1.0.tgz|package/WeexDemo.app', 'WeexPlayground', params);
+  return simrun.ios('weex-devtool', 'iPhone 6', 'https://registry.npm.taobao.org/weex-ios-playground/-/weex-ios-playground-1.2.0.tgz|package/WeexDemo.app', 'wxpage', params);
 };